Refactor: [VoucherType] Enum class에서 바우처 생성과 관련된 코드를 전부 제거 -> VoucherFactory에서 생성을 전담
