+++ src/main/java/com/programmers/voucher/voucher/VoucherType.java
@@ -1,31 +1,15 @@
 package com.programmers.voucher.voucher;
 
+import java.util.Arrays;
 import java.util.List;
-import java.util.UUID;
+
+import static com.programmers.voucher.menu.Message.VOUCHER_INPUT_ERROR_MESSAGE;
 
 
 public enum VoucherType {
-    FixedAmount(List.of("F", "FixedAmount")) {
-        public Voucher createVoucher(long value) {
-            return VoucherFactory.createVoucher(this, value);
-        }
+    FixedAmount(List.of("F", "FixedAmount")),
 
-        @Override
-        public Voucher createVoucher(UUID id, long value) {
-            return VoucherFactory.createVoucher(id, this, value);
-        }
-    },
-
-    PercentDiscount(List.of("P", "PercentDiscount")) {
-        public Voucher createVoucher(long value) {
-            return VoucherFactory.createVoucher(this, value);
-        }
-
-        @Override
-        public Voucher createVoucher(UUID id, long value) {
-            return VoucherFactory.createVoucher(id, this, value);
-        }
-    },
+    PercentDiscount(List.of("P", "PercentDiscount")),
     ;
 
     private final List<String> type;
@@ -38,7 +22,10 @@ public enum VoucherType {
         return type;
     }
 
-    public abstract Voucher createVoucher(long value);
-
-    public abstract Voucher createVoucher(UUID id, long value);
+    public static VoucherType getValidateVoucherType(String inputType) {
+        return Arrays.stream(values())
+                .filter(voucherType -> voucherType.getType().contains(inputType))
+                .findFirst()
+                .orElseThrow(() -> new RuntimeException(VOUCHER_INPUT_ERROR_MESSAGE.getMessage()));
+    }
 }
